requirements.txt
Babel==2.3.4
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
coverage==4.1
Flask==0.10.1
Flask-Babel==0.11.0
Flask-Script==2.0.5
Flask-WTF==0.12
-e git+https://github.com/Fy-Network/fysql@c327448d6748d2bf72eaf16ac761e69649d041a7#egg=fysql
gevent==1.1.1
greenlet==0.4.9
gunicorn==19.6.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.2.0
python-akismet==0.2.3
python-magic==0.4.11
python-memcached==1.58
python-resize-image==1.1.10
pytz==2016.4
redis==2.10.5
requests==2.10.0
six==1.10.0
Unidecode==0.4.19
Werkzeug==0.11.10
WTForms==2.1
